+++ app/__init__.py
@@ -1,0 +1,8 @@
+from flask import Flask, render_template
+from flask_sqlalchemy import SQLAlchemy
+
+app = Flask(__name__)
+app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///monthlysavings.db'
+db = SQLAlchemy(app)
+
+from app import routes
