+++ app/templates/FAQ.html
@@ -1,0 +1,21 @@
+{% extends 'base.html'%}
+
+{% block title%}
+FAQ
+{% endblock%}
+
+{% block content %}
+<!-- CONTENT HERE -->
+
+<p>FAQ Page</p>
+
+<style>
+    body {
+        background-color: #ffffff;
+    }
+
+    .bg-lighter {
+        background-color: #CBCBCB;
+    }
+</style>
+{% endblock%}
