+++ app/templates/base.html
@@ -10,55 +10,38 @@
         integrity="sha384-TX8t27EcRE3e/ihU7zmQxVncDAy5uIKz4rEkgIXeMed4M0jlfIDPvg6uqKI2xXr2" crossorigin="anonymous">
     <!-- responsive meta tag -->
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-    <title>Home</title>
+    <title>
+        {% block title %}
+        {% endblock %}
+    </title>
 </head>
 
 <body>
+    <!-- Navbar here -->
     <nav class="navbar navbar-expand-md navbar-dark bg-lighter">
-        <a class="navbar-brand" href="#">Batteryistic</a>
+        <a class="navbar-brand text-dark font-weight-bold" href="{{ url_for('home_page') }}">Batteryistic</a>
         <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNav">
             <span class="navbar-toggler-icon"></span>
         </button>
         <div class="collapse navbar-collapse" id="navbarNav">
             <ul class="navbar-nav mr-auto">
                 <li class="nav-item active">
-                    <a class="nav-link" href="/home">Home <span class="sr-only">(current)</span></a>
+                    <a class="nav-link text-dark" href="{{ url_for('home_page') }}">Home <span class="sr-only">(current)</span></a>
                 </li>
             </ul>
             <ul class="navbar-nav">
                 <li class="nav-item">
-                    <a class="nav-link" href="/FAQ">FAQ</a>
+                    <a class="nav-link text-dark" href="{{ url_for('FAQ_page') }}">FAQ</a>
                 </li>
-                <!-- <li class="nav-item">
-                    <a class="nav-link" href="/Profile">Profile</a>
-                </li> -->
+                <li class="nav-item">
+                    <a class="nav-link text-dark" href="{{ url_for('Profile_page') }}">Profile</a>
+                </li>
             </ul>
         </div>
     </nav>
-
-    <p>{{welcome}}</p>
-    <div class="container">
-        <table class="table table-hover table-light table-striped table-bordered">
-            <thead>
-                <tr>
-                    <th scope="col">Month/Year</th>
-                    <th scope="col">Current Bill</th>
-                    <th scope="col">Estimated Price</th>
-                    <th scope="col">Savings</th>
-                </tr>
-            </thead>
-            <tbody>
-                {% for item in items %}
-                <tr>
-                    <td>{{item.MonthYear}}</td>
-                    <td>${{item.CurrentBill}}</td>
-                    <td>${{item.EstimatedPrice}}</td>
-                    <td>${{item.Savings}}</td>
-                </tr>
-                {% endfor %}
-            </tbody>
-        </table>
-    </div>
+    <!-- Future Content here -->
+    {% block content %}
+    {% endblock%}
 
     <script src="https://code.jquery.com/jquery-3.5.1.slim.min.js"
         integrity="sha384-DfXdz2htPH0lsSSs5nCTpuj/zy4C+OGpamoFVy38MVBnE+IbbVYUew+OrCXaRkfj"
@@ -66,24 +49,12 @@
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@4.5.3/dist/js/bootstrap.bundle.min.js"
         integrity="sha384-ho+j7jyWK8fNQe+A12Hb8AhRq26LrZ/JpcUGGOn+Y7RsweNrtN/tE3MoK7ZeZDyx"
         crossorigin="anonymous"></script>
-
 </body>
 <style>
     body {
         height: 100%;
         background-color: #ffffff;
         color: black;
-    }
-
-    .bg-lighter {
-        background-color: #CBCBCB;
-    }
-
-    .container {
-        height: 100%;
-        display: flex;
-        justify-content: center;
-        align-items: center;
     }
 </style>
